move cell list typings into view.
